Skip gcc.c-torture/execute/pr60822.c on riscv
The test requires 1.7MB of .bss, but currently only 1MB is supported
before addresses start to wrap around.

There does not appear to be a simple way to skip this for just the
ri5cy board, so the test has been marked as 'skip' for all riscv
targets.

gcc/testsuite/ChangeLog.RISCV:

	* gcc.c-torture/execute/pr60822.c: Skip on riscv as there is not
	enough space on ri5cy for the .bss section.