Stack erase Implementation for RISC-V
Stack erase applied when:

- Epilogue executes
- VLA goes out of scope
- (Corresponding newlib change) Longjmp called

Makes no change to test results, with the exception of
vla-initlist1.C, which relies on the stack remaining unchanged
when a VLA goes out of scope, which is not a sound assumption.

Controlled with the -fstack-erase flag and stack_erase attribute. If
stack erase is not supported on a target it is not accepted.

Add some testing.

Don't inline stack-erase functions.